WTF: Software debayer @ half sensor res  works now!
Before, it was all pink after halfing the sensor resolution with ximea API
("decimation"). Without res halfing, debayering was correct but extremely slow.
I had already experimented with the
format flags, but they were overwritten in the app. Only now, for other purposes,
I stumpled upon the in-app caps, and experimented again:
WITHOUT sensor halfing: bggr is correct
WITH    sensor halfing: grbg is correct
Deducing: Ximeas decimation changes the bayer pattern!
What is even weirder: cv::cuda:demosaicing is unimpressed by those changes,
it works without any hints foth in full and in half res!

Okay, the result is correctly debayerd, but in 2k, high latency and low FPS
(about 20). So, CUDA debayer is still a Must to hit 26 FPS at low latency.

But if everything goes wrong, this is the commit to fall back to!
(Tagged with v0.4.0)