Rendering the demosaiced 4096x3112 buffer easily at full FPS (26.x) and very low latency.
Doing the same hack ximae does (zero copy at the implicit assumption
that the buffer will not be reused before gstreamer is done with it)
vastly improved perdformance. There is a risk, though.
TODO use GstBufferPool to be sure there are no reusal hazards.
But for the prototye, this will have tu suffice.